Extract App layout, toolbars, and dialogs
