+++ src/components/SidebarToggle.tsx
@@ -1,0 +1,24 @@
+import React from 'react';
+import { ChevronLeft, ChevronRight } from 'lucide-react';
+
+type SidebarToggleProps = {
+  show: boolean;
+  leftCollapsed: boolean;
+  onToggle: () => void;
+};
+
+const SidebarToggle = ({ show, leftCollapsed, onToggle }: SidebarToggleProps) => {
+  if (!show) return null;
+  return (
+    <button
+      onClick={onToggle}
+      aria-label={leftCollapsed ? 'Show sidebar' : 'Hide sidebar'}
+      style={{ left: leftCollapsed ? 0 : 224 }}
+      className="no-drag no-press-shift press-fill-effect fixed top-1/2 transform -translate-y-1/2 z-50 bg-white/90 dark:bg-neutral-900/90 border border-border rounded-l-md p-1 shadow-md hover:bg-white dark:hover:bg-neutral-800"
+    >
+      {leftCollapsed ? <ChevronRight className="w-4 h-4" /> : <ChevronLeft className="w-4 h-4" />}
+    </button>
+  );
+};
+
+export default SidebarToggle;
